arcade new logic looking good
