+++ arcade/arcade/templates/{{ toolkit_name }}/pyproject.toml
@@ -6,7 +6,7 @@ authors = ["{{ toolkit_author_name }} <{{ toolkit_author_email }}>"]
 
 [tool.poetry.dependencies]
 python = "^3.10"
-arcade-ai = "^0.1.*"
+arcade-ai = "{{ arcade_version }}"
 
 [tool.poetry.dev-dependencies]
 pytest = "^8.3.0"
