+++ arcade/arcade/templates/{{ toolkit_name }}/{{ package_name }}/tools/my_tool.py
@@ -1,0 +1,10 @@
+from typing import Annotated
+
+from arcade.sdk import tool
+
+
+@tool
+def say_hello(name: Annotated[str, "The name of the person to greet"]) -> str:
+    """Say a greeting!"""
+
+    return "Hello, " + name + "!"
