--- arcade/arcade/templates/{{ toolkit_name }}/{{ toolkit_name }}.py
@@ -1,2 +1,0 @@
-def {{ toolkit_name }}():
-    print("Hello, world!")
